Add New Relic Request Queue Header
My guess here is this is the right data point available to us. Here is the documenation from a New Relic perspective (https://docs.newrelic.com/docs/apm/applications-menu/features/configure-request-queue-reporting/) and here are the docs on request time epoch from API Gateway (https://docs.aws.amazon.com/apigateway/latest/developerguide/api-gateway-mapping-template-reference.html). My hope is that this is the time the requet starts and is prior to cold start or init phase.

NOTE: There is no data for this time epoch in an ALB only integration.